requirements.txt
streamlit
supabase
python-dotenv
streamlit-option-menu
requests
utils
PyPDF2
openai